+++ library-lint/src/main/kotlin/com/evernote/android/state/lint/registry/StateIssueRegistry.kt
@@ -1,9 +1,8 @@
 package com.evernote.android.state.lint.registry
 
 import com.android.tools.lint.client.api.IssueRegistry
-import com.android.tools.lint.detector.api.Issue
-import com.evernote.android.state.lint.detector.StateJavaDetector
+import com.evernote.android.state.lint.detector.AndroidStateDetector
 
 class StateIssueRegistry : IssueRegistry() {
-    override fun getIssues(): List<Issue> = listOf(StateJavaDetector.NonMatchingStateSaverCalls.ISSUE)
+    override val issues = listOf(AndroidStateDetector.ISSUE)
 }