+++ library-lint/src/test/resources/java/ValidActivityOtherMethodKt.kt
@@ -17,13 +17,13 @@ import com.evernote.android.state.StateSaver
 /**
  * @author rwondratschek
  */
-class InvalidActivityNoSaveKt : Activity() {
-    override fun onCreate(savedInstanceState: Bundle) {
+class ValidActivityOtherMethodKt : Activity() {
+    protected override fun onCreate(savedInstanceState: Bundle) {
         super.onCreate(savedInstanceState)
-        StateSaver.restoreInstanceState(this, savedInstanceState)
+        restoreInstanceState()
     }
 
-    override fun onSaveInstanceState(outState: Bundle) {
-        super.onSaveInstanceState(outState)
+    private fun restoreInstanceState() {
+
     }
 }